+++ odoo_score/odoo_shell.py
@@ -37,7 +37,7 @@ import pdb  # pylint: disable=deprecated-module
 standard_library.install_aliases()  # noqa: E402
 
 
-__version__ = '2.0.8'
+__version__ = '2.0.9'
 
 
 MAX_DEEP = 20
@@ -3920,3 +3920,4 @@ def main(cli_args=[]):
 
 if __name__ == "__main__":
     exit(main())
+