+++ clodoo/egg-info/CHANGELOG.rst
@@ -2,6 +2,8 @@
 ~~~~~~~~~~~~~~~~~~~
 
 * [FIX] Rpc with odoo < 10.0
+* [IMP] It does no more depends on os0
+* [IMP] Python 3.6 deprecated
 
 2.0.11 (2024-03-31)
 ~~~~~~~~~~~~~~~~~~~